==> DataWrangling/1_scrape_entities.ipynb <==
import requests
import pickle
import pandas as pd

# Scrape Entities from DPH
The following notebook gets basic information for all the food-service establishments in the Fulton+Deklab area

temp = []
page = 0

# Data for Deklab County
while True:
    r_string = f"https://ga.healthinspections.us/stateofgeorgia/API/index.cfm/search/%7B%22county%22:%22REVLQUxC%22," \
           f"%22permitType%22:%22Rm9vZCBTZXJ2aWNl%22,%22keyword%22:%22%22%7D/{page}"
    req = requests.get(r_string)
    if len(req.json()) > 0:
        temp += req.json()
        page += 1
    else:
        break

temp2 = []
page = 0

# Data for Fulton County
while True:
    r_string = f"https://ga.healthinspections.us/stateofgeorgia/API/index.cfm/search/%7B%22county%22:%22RlVMVE9O%22," \
               f"%22permitType%22:%22Rm9vZCBTZXJ2aWNl%22,%22keyword%22:%22%22%7D/{page}"
    req = requests.get(r_string)
    if len(req.json()) > 0:
        temp2 += req.json()
        page += 1
    else:
        break

temp3 = temp + temp2

# Put Data into Dataframe
## Schema:
- dph_id:str - public health id (for api lookup of prior inspections)
- map_add:str - listed address
- score:int - latest inspection score
- date:str - last inspetion date
- name:str - business name

temp4 = []
for place in temp3:
    try:
        entry = {
            'dph_id': place['id'],
            'map_add': place['mapAddress'].replace(" \r\n", ", "),
            'name': place['name'],
            'score': int(place['columns']['4'].split(':')[-1].strip()),
            'date': place['columns']['5'].split(':')[-1].strip()
        }
        temp4.append(entry)
    except Exception as e:
        print(e, place)

# Save the Raw Records
with open('DPH_entity_raw.pkl', 'wb') as pick:
    pickle.dump(temp4, pick)
# Convert to Dataframe and Save
df2 = pd.DataFrame.from_records(temp4)
df2.to_pickle('DPH_entities_df.pkl')
