==> DataWrangling/2_scape_inspections.ipynb <==
# Scrape Inspection Details
This notebook obtains the inspection details for each entity

import pandas as pd
import requests
from tqdm import tqdm
import pickle

df = pd.read_pickle('DPH_entities_df.pkl')
df

extracted = []
for ids in tqdm(df.dph_id.unique()):
    try:
        req_url = f'https://ga.healthinspections.us/stateofgeorgia/API/index.cfm/inspectionsData/{id}'
        r = requests.get(req_url)
        data = r.json()
    except Exception as e:
        print(e, ids)
        data = None
    d = {
        'id': ids,
        'data': data
    }
    extracted.append(d)

# Save our hard-earned data (over an hour of scraping) to pickle
with open('DPH_inspections_raw.pkl', 'wb') as pick:
    pickle.dump(extracted, pick)


# DB Schema: inspections
The following table saves inspection details on a per-inspection basis. (Each row is an inspection)
- inspection_id: \<int> primary key nullable=false
- dph_entity: 'facilityID' \<str> 64bit encoded nullable=false
- date: \<datetime> nullable=false
- purpose: \<str/categorical> nullable=false
- score: \<int> nullable=false
- inspector: \<str> nullable=false
- printable_path: \<str> nullable=false
- volation_count: \<int> nullable=false
- volations: \<list> nullable=true

from datetime import datetime

rows = []
for entity in extracted:
    entity_id = entity['id']
    entity_data = entity['data']
    for inspection in entity_data:
        inspection_data = inspection['columns']
        inspection_id = inspection['inspectionId']
        if inspection_id != 0:
            violation_count = len(inspection['violations'])
            violation_codes = []
            inspector_notes = []
            if violation_count > 0:
                for k, v in inspection['violations'].items():
                    violation_codes.append(v[0].split(' - ', 1)[0])
                    inspector_notes.append(v[5].replace("Inspector Notes: ", ''))
            try:
                row = {
                    'inspection_id': inspection_id,
                    'dph_id': entity_id,
                    'date': datetime.strptime(inspection_data['0'].replace("Date: ", ''), '%m-%d-%Y').date(),
                    'purpose': inspection_data['1'].replace("Inspection Purpose: ", ''),
                    'score': int(inspection_data['2'].replace("Score: ", '')),
                    'inspector': inspection_data['3'].replace("Inspector: ", ''),
                    'printable_path': inspection['printablePath'],
                    'violation_count': violation_count,
                    'violation_codes': violation_codes,
                    'inspector_notes': inspector_notes
                }
            except Exception as e:
                print(e, inspection, entity)
                break
            rows.append(row)
        else:
            print(entity)
            # These entities don't have inspections on file (perhaps they're new?)

df_inspections = pd.DataFrame.from_records(rows)
# Merge with basic data from part 1.
entity_ids = df[['dph_id', 'name']]
m_df = df_inspections.merge(entity_ids, how='inner', on='dph_id')
m_df.to_pickle('inspection_df.pkl')

m_df

# DB Schema: Violations
The following table saves inspection details on a per-violation basis (each row is a specific violation). In theory,
this will allow every kind of query/aggregation.
- violation_id: \<index>
- inpsection_id: \<int>
- facility_id: \<str>
- violation_desc: \<str>
- points: \<int>
- corrected: \<bool>
- repeat: \<bool>
- notes: \<str>

from datetime import datetime

rows = []
for entity in extracted:
    entity_id = entity['id']
    entity_data = entity['data']
    for inspection in entity_data:
        inspection_data = inspection['columns']
        inspection_id = inspection['inspectionId']
        if inspection_id != 0:
            violation_count = len(inspection['violations'])
            if violation_count > 0:
                for k, v in inspection['violations'].items():
                    violation_raw = v[0].split(' - ', 1)
                    violation_code = violation_raw[0]
                    section_detail = violation_raw[1]
                    inspector_note = v[5].replace("Inspector Notes: ", '')
                    corrected = v[3].replace("Corrected during inspection?: ", '')
                    repeat = v[4].replace("Repeat: ", '')
                    if "Yes" in corrected:
                        corrected = True
                    else:
                        corrected = False
                    if "Yes" in repeat:
                        repeat = True
                    else:
                        repeat = False
                    try:
                        row = {
                            'inspection_id': inspection_id,
                            'dph_id': entity_id,
                            'date': datetime.strptime(inspection_data['0'].replace("Date: ", ''), '%m-%d-%Y').date(),
                            'purpose': inspection_data['1'].replace("Inspection Purpose: ", ''),
                            'score': int(inspection_data['2'].replace("Score: ", '')),
                            'inspector': inspection_data['3'].replace("Inspector: ", ''),
                            'printable_path': inspection['printablePath'],
                            'violation_count': violation_count,
                            'violation_code': violation_code,
                            'section_detail': section_detail,
                            'inspector_note': inspector_note,
                            'corrected': corrected,
                            'repeat': repeat
                        }
                    except Exception as e:
                        print(e, inspection, entity)
                        break
                    rows.append(row)
        else:
            print(entity)

df_violations = pd.DataFrame.from_records(rows)
entity_details = df[['dph_id', 'name', 'types', 'place_id']]
merged_df = df_violations.merge(entity_details, how="inner", on="dph_id")
merged_df.to_pickle('violations_df.pkl')

merged_df

# The most common violations:
merged_df.groupby('violation_code')['violation_count'].sum()
