==> DataWrangling/3_geocoding.ipynb <==
# Geocoding
This notebook takes the entity name & address, and uses google's geocoding API to obtain Lat+Long information

import pandas as pd
import googlemaps
import time
from tqdm import tqdm
import pickle

df = pd.read_pickle('DPH_entities_df.pkl')

gmaps = googlemaps.Client(key='AIzaSyB-rR0pSddf_hka3OETuiAnXx20-ZlOVUE')
geocoded = []

for index, row in tqdm(df.iterrows(), total=df.shape[0]):
    place = row['map_add']
    name = row['name']
    query = f'{name}, {place}'
    try:
        geocode_result = gmaps.geocode(query)
        entry = {
            'index': index,
            'result': geocode_result
        }
        geocoded.append(entry)
    except Exception as e:
        print(e, row)
    time.sleep(30/1000)  # There's a rate limit

# Cache these results, so we don't have to call them again
with open('geocoded_ggl.pkl', 'wb') as pick:
    pickle.dump(geocoded, pick)

# get the top results, because fuckit - good enough
top_res = []
for g in geocoded:
    res = g['result'][0]
    lat = res['geometry']['location']['lat']
    lng = res['geometry']['location']['lng']
    place_id = res['place_id']
    try:
        types = res['types']
        e = {
            'index': g['index'],
            'lat': lat,
            'lng': lng,
            'place_id': place_id,
            'types': types
        }
        top_res.append(e)
    except Exception as e:
        print(e, res)
        break

df_geocoded = pd.DataFrame.from_records(top_res, index='index')
new_df = pd.concat([df, df_geocoded], axis=1)
new_df.to_pickle('geocoded_df.pkl')

new_df
