==> DataWrangling/4_combineComments.ipynb <==
import pandas as pd

df = pd.read_pickle('inspection_df.pkl')
df

## Pandas, Use Aggregate!
- Calculate aggregated values like
    - Average violations over time (Violation Rate)
    - Time in business (years_data)
    - Running average score (int)
    - Combined all the comments (string)

grouped = df.groupby(['dph_id'])

# For each business entity
aggregated_rows = []
for name, group in grouped:
    grouped_comments = []
    grouped_violations = []
    date_delta = group.max(axis=0)['date'] - group.min(axis=0)['date']
    data_date_range = date_delta.days/365
    lifetime_violations = group['violation_count'].sum()
    lifetime_avg = group['score'].mean()
    for index, inspection in group.iterrows():
        if len(inspection['inspector_notes']) > 0:
            grouped_comments += inspection['inspector_notes']
        if len(inspection['violation_codes']) > 0:
            grouped_violations += inspection['violation_codes']

    # Calculate violation rate: if operating less than 1 year, violation rate = lifetime violations
    # else lifetime rate = violations/years in operation
    if data_date_range <= 1.0:
        lifetime_rate = lifetime_violations
    else:
        lifetime_rate = lifetime_violations/data_date_range

    row = {
        'dph_id': name,
        'violations': grouped_violations,
        'years_data': data_date_range,
        'lifetime_count': lifetime_violations,
        'lifetime_comments': ' '.join(grouped_comments).replace('\r\n', ''),
        'lifetime_avg': lifetime_avg,
        'violation_rate': lifetime_rate
    }
    aggregated_rows.append(row)

agg_df = pd.DataFrame.from_records(aggregated_rows)
agg_df.to_pickle('entity_aggregation.pkl')
agg_df
