==> DataWrangling/6_getYelpData.ipynb <==
import pandas as pd
import time
from tqdm import tqdm
from yelpapi import YelpAPI

# Get Yelp Business data
Given that these are all the places with reviews, get their information from Yelp

df1 = pd.read_pickle('all_yelp_ids.pkl')
df2 = pd.read_pickle('geocoded_df.pkl')
df3 = df1.merge(df2, how='inner', on='dph_id')

api_key = "qOWiNCp8Sblg84q0uiPRXiksxMk2DzMvMRnnSI29PxS-xv9sG9e4yxJ-lPHFaMHaacOJ8uBOZX4T1rY3fhnZTzzy4ZW2m_vT-WhnYgix5RSSc4VnFQw0-lQmQVQxZHYx"

responses = []
frame = df3
for index, entity in tqdm(frame.iterrows(), total=frame.shape[0]):
    try:
        with YelpAPI(api_key) as yelp_api:
            response = yelp_api.business_query(id=entity['yelp_id'])
        row = {
            'dph_id': entity['dph_id'],
            'result': response
        }
    except Exception as e:
        row = {
            'dph_id': entity['dph_id'],
            'result': []
        }
        print(e)
    responses.append(row)
    time.sleep(20/1000)

filt_resp = []
for r in responses:
    if len(r['result']) != 0:
        filt_resp.append(r)

filt_resp = pd.DataFrame.from_records(filt_resp)
filt_resp.to_pickle('yelp_all_data.pkl')
filt_resp

# Serialize these responses we're interested in
serialized = []

for index, resp in filt_resp.iterrows():
    r = resp['result']
    price = len(r.get('price', ''))
    s = {
        'dph_id': resp['dph_id'],
        'yelp_id': r['id'],
        'rating': r['rating'],
        'review_count': r['review_count'],
        'img_url': r['image_url'],
        'yelp_name': r['name'],
        'price': price
    }
    serialized.append(s)

serialized = pd.DataFrame.from_records(serialized)
serialized

agg_df = pd.read_pickle('entity_aggregation.pkl')

yelp_df = serialized.merge(agg_df, how='inner', on='dph_id')

yelp_merged_df = yelp_df.merge(df2, how='inner', on='dph_id')
yelp_merged_df.to_pickle('final_yelpd_df.pkl')
yelp_merged_df
