==> DataWrangling/8_makeGeoJSON.ipynb <==
# Make GeoJSON
This notebook takes in the output of our data exploration, and compiles it to a GeoJSON, which leaflet can plot

import pandas as pd
import json
survivors = pd.read_pickle('final_dataset_w_yelp.pkl')

##  Feature Collection
We wish to have the following schema for GeoJSON Feature Collections:
```
{
    'type':'FeatureCollection',
    'features':[{
        'type': 'Feature'
        'geometry': {'type':'Point', 'coordinates':[lat, long]}
        'properties': {'prop1':val, ...}
    }...]
}
```

def row2feat(row):
    feat = {
        'type': 'Feature',
        'geometry': {'type': 'Point', 'coordinates': [row['lng'], row['lat']]},
        'properties': {
            'score': row['score'],
            'date': row['date'],
            'violation_rate': row['violation_rate'],
            'lifetime_count': row['lifetime_count'],
            'lifetime_avg': row['lifetime_avg'],
            'rating': row['rating'],
            'review_count': row['review_count'],
            'years_data': row['years_data'],
            'pests': row['pests'],
            'personnel': row['personnel'],
            'sanitation': row['sanitation'],
            'temperature': row['temperature'],
            'hazard': row['hazard'],
            'img_url': row['img_url'],
            'name': row['name'],
            'addr': row['map_add'],
            'id': row['dph_id']
        }
    }
    return feat

# Get dataframe into json
features = []
for index, row in survivors.iterrows():
    features.append(row2feat(row))

geoJSON = {
    'type': 'FeatureCollection',
    'features': features
}

with open('final_collection.json', 'w') as f:
    json.dump(geoJSON, f)
